requirements.txt
faiss-cpu
transformers
torch
numpy
together
requests
certifi
